Chirp now runs in the background
